# 03-Azure/01-04-AI/01_Agentic_SDLC/src/api-ts/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/cart/{branchId}": {
            "get": {
                "summary": "Get the active cart for a branch",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "branchId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Active or empty cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cart/{branchId}/items": {
            "post": {
                "summary": "Add a product to the active cart",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "branchId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated cart"
                    }
                }
            }
        },
        "/api/cart/{branchId}/items/{itemId}": {
            "put": {
                "summary": "Set a cart item's quantity",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "branchId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "path",
                        "name": "itemId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCartItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Remove an item from the cart",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "branchId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "path",
                        "name": "itemId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item removed"
                    }
                }
            }
        },
        "/api/cart/{branchId}/payment/setup": {
            "get": {
                "summary": "Get Google Authenticator-compatible setup details for simulated payment",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "branchId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TOTP setup details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentSetup"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cart/{branchId}/checkout": {
            "post": {
                "summary": "Complete a simulated payment and convert the cart into an order",
                "tags": [
                    "Cart"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "branchId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckoutRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment approved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentConfirmation"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Simulated payment declined"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CartItem": {
                "type": "object",
                "properties": {
                    "orderDetailId": {
                        "type": "integer"
                    },
                    "productId": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "imgName": {
                        "type": "string"
                    },
                    "unit": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "unitPrice": {
                        "type": "number",
                        "format": "float"
                    },
                    "lineTotal": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
            "Cart": {
                "type": "object",
                "properties": {
                    "orderId": {
                        "type": "integer",
                        "nullable": true
                    },
                    "branchId": {
                        "type": "integer"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CartItem"
                        }
                    },
                    "total": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
            "AddCartItemRequest": {
                "type": "object",
                "required": [
                    "productId",
                    "quantity"
                ],
                "properties": {
                    "productId": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "minimum": 1
                    }
                }
            },
            "UpdateCartItemRequest": {
                "type": "object",
                "required": [
                    "quantity"
                ],
                "properties": {
                    "quantity": {
                        "type": "integer",
                        "minimum": 1
                    }
                }
            },
...
            "PaymentSetup": {
                "type": "object",
                "properties": {
                    "issuer": {
                        "type": "string"
                    },
                    "accountName": {
                        "type": "string"
                    },
                    "qrCodeDataUrl": {
                        "type": "string"
                    },
                    "manualEntryKey": {
                        "type": "string"
                    }
                }
            },
            "CheckoutRequest": {
                "type": "object",
                "required": [
                    "cardholderName",
                    "paymentMethod",
                    "otp"
                ],
                "properties": {
                    "cardholderName": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 100
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "demo-card-approved",
                            "demo-card-declined"
                        ]
                    },
                    "otp": {
                        "type": "string",
                        "pattern": "^\\\\d{6}$"
                    }
                }
            },
            "PaymentConfirmation": {
                "type": "object",
                "properties": {
                    "paymentAttemptId": {
                        "type": "integer"
                    },
                    "orderId": {
                        "type": "integer"
                    },
                    "branchId": {
                        "type": "integer"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "providerReference": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "approved"
                        ]
                    },
                    "orderStatus": {
                        "type": "string",
                        "enum": [
                            "pending"
                        ]
                    },
                    "paidAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    },
...
}